manager: Initiate support bundle now clean existing request automatically
Make it easier for UI to repeat request if needed.

The downside is, if multiple users request the bundle at the same time,
some of the requests will be denied and user would need to retry.
